[chore][fileconsumer] Establish internal util package
