+++ pkg/stanza/fileconsumer/reader_test.go
@@ -222,27 +222,6 @@ func readToken(t *testing.T, c chan *emitParams) []byte {
 	return nil
 }
 
-func TestMapCopy(t *testing.T) {
-	initMap := map[string]any{
-		"mapVal": map[string]any{
-			"nestedVal": "value1",
-		},
-		"intVal": 1,
-		"strVal": "OrigStr",
-	}
-
-	copyMap := mapCopy(initMap)
-	// Mutate values on the copied map
-	copyMap["mapVal"].(map[string]any)["nestedVal"] = "overwrittenValue"
-	copyMap["intVal"] = 2
-	copyMap["strVal"] = "CopyString"
-
-	// Assert that the original map should have the same values
-	assert.Equal(t, "value1", initMap["mapVal"].(map[string]any)["nestedVal"])
-	assert.Equal(t, 1, initMap["intVal"])
-	assert.Equal(t, "OrigStr", initMap["strVal"])
-}
-
 func TestEncodingDecode(t *testing.T) {
 	testFile := openTemp(t, t.TempDir())
 	testToken := tokenWithLength(2 * fingerprint.DefaultSize)
